+++ src/Sulu/Bundle/ContactBundle/Resources/public/js/main.js
@@ -41,6 +41,9 @@ define(['config', 'aura_extensions/iban'], function(Config, IbanExtension) {
 
             var sandbox = app.sandbox;
 
+            sandbox.urlManager.setUrl('contact', 'contacts/contacts/edit:<%= id %>/details');
+            sandbox.urlManager.setUrl('account', 'contacts/accounts/edit:<%= id %>/details');
+
             Config.set('sulucontact.components.autocomplete.default.contact', {
                 remoteUrl: '/admin/api/contacts?searchFields=id,fullName&flat=true&fields=id,fullName&limit=25',
                 getParameter: 'search',