+++ src/Sulu/Bundle/MediaBundle/Media/Manager/MediaManager.php
@@ -531,7 +531,7 @@ class MediaManager implements MediaManagerInterface
     private function buildData($uploadedFile, $data, $user)
     {
         if (!($uploadedFile instanceof UploadedFile)) {
-            throw new InvalidFileException('given uploadfile is not of instance UploadFile');
+            throw new InvalidFileException('Given uploaded file is not of instance UploadedFile');
         }
 
         $this->validator->validate($uploadedFile);